abaqus/portfolio/requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
Django==5.2
django-rest-framework==0.1.0
djangorestframework==3.16.0
executing==2.2.0
fastexcel==0.13.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
polars==1.27.1
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pyzmq==26.4.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
wcwidth==0.2.13
